Fix detection of default remote branch when it's not specified at CLI
