+++ change/beachball-1efd0dd6-cc5b-4ddd-a70a-b32189190941.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Fix detection of default remote branch when it's not specified at CLI",
+  "packageName": "beachball",
+  "email": "elcraig@microsoft.com",
+  "dependentChangeType": "patch"
+}
